requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
alembic
pydantic
pydantic_settings
python-jose[cryptography]
passlib[bcrypt]
redis
python-dotenv
email-validator
python-multipart